fix(settings): widen pagination offset to i64 + un-garble the pager doc
Code-review follow-ups on the BUNYIP-177 pagination.

1. offset overflow: list_sessions / list_trusted_devices computed `offset = (page - 1) * per_page` as i32 with no upper bound on `page` (only `.max(1)`). A crafted `?page=2147483647&per_page=2` overflowed i32 - a negative OFFSET in release (Postgres rejects it -> 500) or a panic in debug. Compute the offset as i64 (per_page stays clamped to 1..=100) and widen the repo `offset` params to i64 so a large page just returns an empty page.

2. doc comment: the new settings_pager helper was inserted into the middle of sessions_card_body's two-part doc comment, so settings_pager's doc started with "Body of the Active Sessions card..." and sessions_card_body lost its summary. Restore each function's own doc.

clippy --all-targets -D warnings and fmt --check clean.

#BUNYIP-177

+++ bunyip-api/src/handlers/user.rs
@@ -270,7 +270,9 @@ pub async fn list_sessions(
     let request_id = get_request_id(&req);
     let page = query.page.unwrap_or(1).max(1);
     let per_page = query.per_page.unwrap_or(20).clamp(1, 100);
-    let offset = (page - 1) * per_page;
+    // i64 so a crafted large ?page= cannot overflow the offset (a negative
+    // OFFSET would 500). per_page is already clamped to 1..=100 (BUNYIP-177).
+    let offset = (page as i64 - 1) * per_page as i64;
 
     // Hash the caller's refresh-token cookie (if present) to match it to its
     // stored row without exposing any hashes to the client.
@@ -674,7 +676,9 @@ pub async fn list_trusted_devices(
     let request_id = get_request_id(&req);
     let page = query.page.unwrap_or(1).max(1);
     let per_page = query.per_page.unwrap_or(20).clamp(1, 100);
-    let offset = (page - 1) * per_page;
+    // i64 so a crafted large ?page= cannot overflow the offset (a negative
+    // OFFSET would 500). per_page is already clamped to 1..=100 (BUNYIP-177).
+    let offset = (page as i64 - 1) * per_page as i64;
 
     let devices =
         TrustedDeviceRepository::find_user_devices_paginated(&pool, user.0.sub, per_page, offset)
